third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGRadialGradientElement-svgdom-cx-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'cx' property of the SVGRadialGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS radialGradientElement.cx.baseVal.valueAsString is "0%"
PASS radialGradientElement.cx.baseVal.valueAsString is "150%"
PASS successfullyParsed is true

TEST COMPLETE
